fix iou calc function

+++ utils/metrics.py
@@ -8,7 +8,7 @@ import numpy as np
 
 def compute_iou(outputs, targets):
     intersection = (outputs & targets).float().sum((1, 2))
-    union = (outputs | targets).float().sum(1, 2)
+    union = (outputs | targets).sum(1, 2)
 
     iou = (intersection + SMOOTH) / (union + SMOOTH)
